refactor(github, onboarding): 웹훅 예외 및 검증 어댑터 분리
