+++ backend/src/main/java/com/ssafy/dash/github/infrastructure/webhook/WebhookSignatureValidatorImpl.java
@@ -1,4 +1,4 @@
-package com.ssafy.dash.github.infrastructure.client;
+package com.ssafy.dash.github.infrastructure.webhook;
 
 import java.nio.charset.StandardCharsets;
 
@@ -9,16 +9,16 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-import org.springframework.web.server.ResponseStatusException;
-import org.springframework.http.HttpStatus;
 
 import com.ssafy.dash.github.config.GitHubWebhookProperties;
 import com.ssafy.dash.github.domain.WebhookSignatureValidator;
+import com.ssafy.dash.github.domain.exception.GitHubWebhookException;
 
 @Component
 public class WebhookSignatureValidatorImpl implements WebhookSignatureValidator {
 
     private static final Logger log = LoggerFactory.getLogger(WebhookSignatureValidatorImpl.class);
+
     private final GitHubWebhookProperties properties;
 
     public WebhookSignatureValidatorImpl(GitHubWebhookProperties properties) {
@@ -47,7 +47,7 @@ public class WebhookSignatureValidatorImpl implements WebhookSignatureValidator 
             return toHex(digest);
         } catch (Exception ex) {
             log.error("Failed to verify GitHub webhook signature", ex);
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "웹훅 서명 검증 중 오류가 발생했습니다.");
+            throw new GitHubWebhookException("웹훅 서명 검증 중 오류가 발생했습니다.", ex);
         }
     }
 
@@ -71,5 +71,4 @@ public class WebhookSignatureValidatorImpl implements WebhookSignatureValidator 
         }
         return result == 0;
     }
-    
 }
