+++ backend/src/main/java/com/ssafy/dash/github/presentation/GitHubWebhookController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ssafy.dash.github.domain.WebhookSignatureValidator;
+import com.ssafy.dash.github.domain.exception.GitHubWebhookException;
 
 @RestController
 @RequestMapping("/api/webhooks/github")
@@ -31,9 +32,15 @@ public class GitHubWebhookController {
             @RequestBody(required = false) String payload) {
         String body = payload == null ? "" : payload;
 
-        if (!signatureValidator.isValid(signature, body)) {
-            log.warn("Invalid GitHub webhook signature. Event: {}, Signature: {}", event, signature);
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("GitHub webhook 서명이 유효하지 않습니다.");
+        try {
+            if (!signatureValidator.isValid(signature, body)) {
+                log.warn("Invalid GitHub webhook signature. Event: {}, Signature: {}", event, signature);
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("GitHub webhook 서명이 유효하지 않습니다.");
+            }
+        } catch (GitHubWebhookException ex) {
+            log.error("GitHub webhook signature validation failed", ex);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("GitHub 웹훅 처리 중 오류가 발생했습니다.");
         }
 
         if ("ping".equalsIgnoreCase(event)) {
